requirements.txt
Django==1.9.12
djangorestframework==3.5.3
evernote==1.25.0
httplib2==0.9.2
Markdown==2.6.7
oauth2==1.9.0.post1
oauthlib==2.0.1
py==1.4.32
requests==2.12.4
requests-oauthlib==0.7.0
